Added support for DAC control of Vref in generic stepper driver
Extended the generic stepper driver class to be able to change the value of Vref for stepper drivers that can be controlled as such. Needs a DAC output on the MCU, tested on the ESP32 with two motors. Added dacWriteEx macro as well for that job. I added an extra parameter in stepDirDriverPins and stepDirrDriverSettings for the pin and Vref divider values.